Refactor hmc5883l_compass blocks into separate module.
