+++ src/modules/editor.js
@@ -48,6 +48,7 @@ import './blockly/generators/propc/sensors/color_pal';
 import './blockly/generators/propc/sensors/dht22';
 import './blockly/generators/propc/sensors/fingerprint_scanner';
 import './blockly/generators/propc/sensors/gps';
+import './blockly/generators/propc/sensors/hmc5883l_compass';
 import './blockly/generators/propc/sensors/joystick';
 import './blockly/generators/propc/sensors/mma7455';
 import './blockly/generators/propc/sensors/mx2125';
